@@ -2,5 +2,5 @@ package com.andersenlab.assesment.dto;
 
 import java.util.List;
 
-public record OwnerDogLinkDto(List<String> dogs) {
+public record OwnerDogLinkDto(List<Integer> dogIds) {
 }
