@@ -1,8 +1,8 @@
 package com.andersenlab.assesment.entity;
 
+import jakarta.persistence.*;
 import lombok.*;
 
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,38 +11,19 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@EqualsAndHashCode(of = { "firstName", "lastName", "age", "city"})
-@ToString(exclude = "dogs")
-@NamedEntityGraph(name = "owner-graph", attributeNodes = {@NamedAttributeNode("dogs")})
+//@NamedEntityGraph(name = "owner-graph", attributeNodes = {@NamedAttributeNode("dogs")})
 public class Owner {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @ToString.Exclude
     private Integer id;
-
     private String firstName;
-
     private String lastName;
-
     private Integer age;
-
     private String city;
-
-    @ManyToMany
-    @JoinTable(
-            name = "dog_owner",
-            joinColumns = @JoinColumn(name = "owner_id"),
-            inverseJoinColumns = @JoinColumn(name = "dog_id")
-    )
+    private String email;
+    @OneToMany(mappedBy = "owner", cascade = {CascadeType.REMOVE})
+    @ToString.Exclude
     private List<Dog> dogs = new ArrayList<>();
-
-    public void addDogs(List<Dog> dogs) {
-        this.dogs.addAll(dogs);
-        dogs.forEach(dog -> dog.getOwners().add(this));
-    }
-
-    public void removeDogs(List<Dog> dogs) {
-        this.dogs.removeAll(dogs);
-        dogs.forEach(dog -> dog.getOwners().remove(this));
-    }
 }
