@@ -13,7 +13,11 @@ public enum ErrorCode {
     ERR002(ErrorMessage.ERR002_MESSAGE),
     ERR003(ErrorMessage.ERR003_MESSAGE),
     ERR004(ErrorMessage.ERR004_MESSAGE),
-    ERR005(ErrorMessage.ERR005_MESSAGE);
+    ERR005(ErrorMessage.ERR005_MESSAGE),
+    ERR006(ErrorMessage.ERR006_MESSAGE),
+    ERR007(ErrorMessage.ERR007_MESSAGE),
+    ERR008(ErrorMessage.ERR008_MESSAGE),
+    ERR009(ErrorMessage.ERR009_MESSAGE);
 
     private final String message;
 
